Moved some interfaces around.
