+++ subprojects/platform-native/src/main/java/org/gradle/nativeplatform/toolchain/internal/NativeLanguageTools.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.nativeplatform.toolchain.internal.msvcpp;
+package org.gradle.nativeplatform.toolchain.internal;
 
 import java.io.File;
