+++ subprojects/platform-native/src/main/java/org/gradle/nativeplatform/toolchain/internal/msvcpp/AbstractWindowsKitComponentLocator.java
@@ -23,6 +23,7 @@ import net.rubygrapefruit.platform.MissingRegistryEntryException;
 import net.rubygrapefruit.platform.WindowsRegistry;
 import org.gradle.api.Transformer;
 import org.gradle.internal.FileUtils;
+import org.gradle.platform.base.internal.toolchain.SearchResult;
 import org.gradle.util.CollectionUtils;
 import org.gradle.util.VersionNumber;
 import org.slf4j.Logger;
