+++ subprojects/platform-native/src/main/java/org/gradle/nativeplatform/toolchain/internal/msvcpp/PlatformWindowsSdk.java
@@ -16,5 +16,7 @@
 
 package org.gradle.nativeplatform.toolchain.internal.msvcpp;
 
+import org.gradle.nativeplatform.toolchain.internal.SystemLibraries;
+
 public interface PlatformWindowsSdk extends WindowsTools, SystemLibraries {
 }
